refactor: extract useHealth hook, delete dead StatusBar
Layout now consumes a useHealth() hook instead of inlining the polling
useEffect. Old StatusBar is fully orphaned after the shell migration, so
it's removed.
